Change database script location and add insert example script
