+++ MainDirectory/database/drop_tables.sql
